refactor(gpu_circuit_prover): type domain-size log2 values as u32
The log2 authorities are already u32 (`DOMAIN_SIZE_LOG2` trait consts,
`inits_and_teardowns::TRACE_LEN_LOG2`), and `trailing_zeros()` — the
conventional source of such values — returns u32, as do the GPU-side log
params that consume them. Typing `get_domain_size_log2()` (all five enums)
and the `*_DOMAIN_SIZE_LOG2` consts as u32 removes the `as usize` casts at
the sources and the `as u32` re-casts at GPU-param call sites; the remaining
`as usize` casts sit only at genuine CPU-API boundaries (the cs compiler's
`trace_len_log2: usize`, `example_configs`, `collect_inits_and_teardowns_
into_columns`). Net: fewer casts, and the type flows forward from the
authority instead of round-tripping.

Values unchanged; full test-build + circuit_type unit tests green.

+++ gpu/circuit_prover/src/prover/config.rs
@@ -67,7 +67,7 @@ fn config_for_supported_level(
             )
         }
     };
-    config_for_security_level_under_pessimistic_conjecture(schedule_log_2, security_level)
+    config_for_security_level_under_pessimistic_conjecture(schedule_log_2 as usize, security_level)
 }
 
 /// PoW bit count for the lookup challenges (`lookup_alpha`, `lookup_additive`),
